Added step by step instructions for use.
I used this module to back up my own Flickr account and I had a lot of
issues struggling with the auth_token and frob stuff, so when I figured
it out I thought I'd better document it.

Included is a script for fetching the authorization URL for your app. I
called it net-flickr-backup-get-auth-token to avoid coliding with other
binaries installed on the system.

Finally, I renamed backup.pl to net-flickr-backup-do-backup for the same
reason mentioned above, and for consistency.